Move work from isni_matching_modified.py to isni_matching.py
